Refactor: tutorial center graph change
